+++ packages/util-crypto/src/base58/bs58.ts
@@ -18,7 +18,7 @@ const config = {
  * @description
  * Validates that the supplied value is valid base58, throwing exceptions if not
  */
-export const base58Validate = createValidate(config);
+export const base58Validate = /*#__PURE__*/ createValidate(config);
 
 /**
  * @name base58Decode
@@ -26,7 +26,7 @@ export const base58Validate = createValidate(config);
  * @description
  * From the provided input, decode the base58 and return the result as an `Uint8Array`.
  */
-export const base58Decode = createDecode(config, base58Validate);
+export const base58Decode = /*#__PURE__*/ createDecode(config, base58Validate);
 
 /**
 * @name base58Encode
@@ -34,10 +34,10 @@ export const base58Decode = createDecode(config, base58Validate);
 * @description
 * From the provided input, create the base58 and return the result as a string.
 */
-export const base58Encode = createEncode(config);
+export const base58Encode = /*#__PURE__*/ createEncode(config);
 
 /**
 * @name isBase58
 * @description Checks if the input is in base58, returning true/false
 */
-export const isBase58 = createIs(base58Validate);
+export const isBase58 = /*#__PURE__*/ createIs(base58Validate);