+++ packages/util/src/bi/min.ts
@@ -7,10 +7,10 @@ import { createCmp } from './helpers';
  * @name nMax
  * @summary Finds and returns the highest value in an array of bigint.
  */
-export const nMax = createCmp<bigint>((a, b) => a > b);
+export const nMax = /*#__PURE__*/ createCmp<bigint>((a, b) => a > b);
 
 /**
  * @name nMin
  * @summary Finds and returns the lowest value in an array of bigint.
  */
-export const nMin = createCmp<bigint>((a, b) => a < b);
+export const nMin = /*#__PURE__*/ createCmp<bigint>((a, b) => a < b);