+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes:
 - `has{Buffer, Process}` now checks on `globalThis` (helps bundlers with auto-injection)
 - Add `/*#__PURE__*/` annotations for specific `export const something = someFunction(...)`
 - Upgrade dependencies to latest stable versions
+- Upgrade to `@polkadot/wasm-crypto` 6.4.1
 
 
 ## 10.1.14 Nov 27, 2022